[MISC,TEST] Move fast_istreambuf_iterator into io/stream/iterator
