+++ test/unit/io/stream/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_test(fast_istreambuf_iterator_test.cpp)
